Adds NodeType and NodeConfig to NodeResponse model
This commit adds the NodeType and NodeConfig fields to the NodeResponse model in order to provide additional information about the node. These fields will be used to store the type and configuration of the node. This change improves the completeness of the NodeResponse model and allows for better representation of node data.